requirements.txt
numpy
scipy
h5py
tqdm
matplotlib
opencv-python

torch
torchvision
diffusers
accelerate

gymnasium
mani_skill
sapien
zarr

pyzmq
python-can
piper_sdk
